Move MockCommandRunner to helpers::test_utils to share the implementation between modules
